dummy-client/requirements.txt
psycopg2-binary==2.9.6
prometheus-client==0.17.1 